requirements.txt
Pyrofork
tgcrypto
motor
marshmallow
umongo
pymongo
pytz
cinemagoer
shortzy
aiohttp
aiofiles
telegraph
speedtest-cli
uvloop
python-dotenv
HorridAPI
tinydb
langdetect
spacy